Migrating from services to dedicated API files
