+++ app/src/components/editor/ContentView.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Text, Center } from '@mantine/core';
 import Editor from './Editor';
 import MarkdownPreview from './MarkdownPreview';
-import { getFileUrl } from '../../services/api';
+import { getFileUrl } from '../../api/notes';
 import { isImageFile } from '../../utils/fileHelpers';
 
 const ContentView = ({
