+++ app/src/hooks/useFileOperations.js
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { notifications } from '@mantine/notifications';
-import { saveFileContent, deleteFile } from '../services/api';
+import { saveFileContent, deleteFile } from '../api/notes';
 import { useWorkspace } from '../contexts/WorkspaceContext';
 import { useGitOperations } from './useGitOperations';
