+++ app/src/hooks/useGitOperations.js
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { notifications } from '@mantine/notifications';
-import { pullChanges, commitAndPush } from '../services/api';
+import { pullChanges, commitAndPush } from '../api/notes';
 import { useWorkspace } from '../contexts/WorkspaceContext';
 
 export const useGitOperations = () => {
